Fixing 2nd door issue
There is no need for a new function to open the 2nd door. Use the openTheDoor() with the parameter needed.

There is no need for a new keyPress event. I assume that the reason why you could not open the doors is because you could not collect the keys - the new keyPress event was bypassing the other one understanding the player's movement. So, in the previous event I 've added an if to check if the number you need to press is pressed.